Added Middleware for user auth notes

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,14 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+    const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -14,7 +18,9 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
     try {
+        assertIsDefined(authenticatedUserId);
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid note id");
         }
@@ -23,7 +29,9 @@ export const getNote: RequestHandler = async (req, res, next) => {
         if (!note){
             throw createHttpError(404, "Note not found");
         }
-
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "Not authorized to access this note");
+        } 
         res.status(200).json(note);
     } catch (error) {
         next(error);
@@ -38,11 +46,15 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody,unknown> = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
+    
+    const authenticatedUserId = req.session.userId;
     try {
+        assertIsDefined(authenticatedUserId);
         if (!title){
             throw createHttpError(400, "Title is required");
         }
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
         });
@@ -65,7 +77,10 @@ export const updateNote: RequestHandler<updatedNoteParams, unknown, UpdateNoteBo
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    
+    const authenticatedUserId = req.session.userId;
     try {
+        assertIsDefined(authenticatedUserId);
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid note id");
         }
@@ -77,6 +92,10 @@ export const updateNote: RequestHandler<updatedNoteParams, unknown, UpdateNoteBo
         if (!updatedNote){
             throw createHttpError(404, "Note not found");
         }
+        
+        if(!updatedNote.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "Not authorized to access this note");
+        } 
 
         updatedNote.title = newTitle;
         updatedNote.text = newText;
@@ -89,7 +108,10 @@ export const updateNote: RequestHandler<updatedNoteParams, unknown, UpdateNoteBo
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    
+    const authenticatedUserId = req.session.userId;
     try {
+        assertIsDefined(authenticatedUserId);
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid note id");
         }
@@ -98,7 +120,9 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
         if (!note){
             throw createHttpError(404, "Note not found");
         }
-
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "Not authorized to access this note");
+        } 
         await note.deleteOne();
 
         res.sendStatus(204);
